# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Docs",
        "description": "Dokumentasi API",
        "version": "1.0.0"
    },
    "paths": {
        "/api/login": {
            "post": {
                "tags": [
                    "APP MOBILE HIMAKOM"
                ],
                "description": "Login user",
                "operationId": "222b2625838e3c57e017987f73598fda",
                "requestBody": {
                    "description": "User credentials",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "description": "User email",
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "password": {
                                        "description": "User password",
                                        "type": "string",
                                        "format": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "422": {
                        "description": "The given data was invalid.",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "500": {
                        "description": "Login failed.",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
        "/api/logout": {
            "post": {
                "tags": [
                    "APP MOBILE HIMAKOM"
                ],
                "description": "Logout user",
                "operationId": "ad65cbbd4e9f201619eae184a5961a98",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "500": {
                        "description": "Logout failed.",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated.",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/cabinet": {
            "get": {
                "tags": [
                    "APP MOBILE HIMAKOM"
                ],
                "description": "Return data cabinet",
                "operationId": "f7cb2da1822254562a99b3f443c77fa7",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "500": {
                        "description": "Failed to get cabinet data.",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated.",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/events": {
            "get": {
                "tags": [
                    "APP MOBILE HIMAKOM"
                ],
                "description": "Return data events",
                "operationId": "6c05b2a267d0ea183741e1c32f8c6792",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "500": {
                        "description": "Failed to get events data.",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated.",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/notifications": {
            "get": {
                "tags": [
                    "APP MOBILE HIMAKOM"
                ],
                "description": "Return data notifications",
                "operationId": "3ff1df38374626020c33eab48679473a",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "500": {
                        "description": "Failed to update device token.",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated.",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/notifications/{notification}/read": {
            "get": {
                "tags": [
                    "APP MOBILE HIMAKOM"
                ],
                "description": "Update notification to read",
                "operationId": "b059d74e6ce949de8f324e4e44715d61",
                "parameters": [
                    {
                        "name": "notification",
                        "in": "path",
                        "description": "Notification ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "500": {
                        "description": "Failed to update device token.",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated.",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/user": {
            "get": {
                "tags": [
                    "APP MOBILE HIMAKOM"
                ],
                "description": "Return data user profile",
                "operationId": "fd8e4e27fd1a9e98a6638cec7e2126ca",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "500": {
                        "description": "Failed to get user data.",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated.",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/user/device-token": {
            "put": {
                "tags": [
                    "APP MOBILE HIMAKOM"
                ],
                "description": "Update device token when user login",
                "operationId": "36eed3f481fd480fd93d0f3e007651f2",
                "requestBody": {
                    "description": "Update device token when user login",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "device_token"
                                ],
                                "properties": {
                                    "device_token": {
                                        "type": "string",
                                        "example": "device_token"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "500": {
                        "description": "Failed to update device token.",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated.",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "name": "Authorization Control",
                "scheme": "bearer"
            }
        }
    }
}